server/requirements.txt
# server/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyMuPDF
python-dotenv
httpx
google-genai
python-multipart
celery[redis]
redis
pdfplumber
pytesseract
Pillow
sentence-transformers
reportlab
sentry-sdk[fastapi]
